optional xattributes

+++ client/src/fuse/unix.rs
@@ -161,16 +161,20 @@ impl PathFilesystem for Fs {
         name: &OsStr,
         size: u32,
     ) -> FuseResult<ReplyXAttr> {
-        let path = Path::new(path);
-        let name = name.to_str().ok_or_else(|| FuseError::InvalidInput("Attributes name is not valid UTF-8".to_string()))?;;
-        let xattr = self.fs.get_x_attributes(path, name).await?;
-        if size == 0{
-            return Ok(ReplyXAttr::Size(xattr.len() as u32))
+        if self.fs.use_xattributes() {
+            let path = Path::new(path);
+            let name = name.to_str().ok_or_else(|| FuseError::InvalidInput("Attributes name is not valid UTF-8".to_string()))?;;
+            let xattr = self.fs.get_x_attributes(path, name).await?;
+            if size == 0 {
+                return Ok(ReplyXAttr::Size(xattr.len() as u32))
+            }
+            if size < xattr.len() as u32 {
+                return Err(libc::ERANGE.into());
+            }
+            Ok(ReplyXAttr::Data(Bytes::from(xattr)))
+        } else {
+            Err(FuseError::Unsupported("getxattr".to_string()).into())
         }
-        if size < xattr.len() as u32 {
-            return Err(libc::ERANGE.into());
-        }
-        Ok(ReplyXAttr::Data(Bytes::from(xattr)))
     }
 
     /// set an extended attribute.
@@ -184,39 +188,51 @@ impl PathFilesystem for Fs {
         flags: u32,
         position: u32,
     ) -> FuseResult<()> {
-        let path = Path::new(path);
-        let name = name.to_str().ok_or_else(|| FuseError::InvalidInput("Attributes name is not valid UTF-8".to_string()))?;;
-        self.fs.set_x_attributes(path, name, value, flags, position).await;
-        Ok(())
+        if self.fs.use_xattributes() {
+            let path = Path::new(path);
+            let name = name.to_str().ok_or_else(|| FuseError::InvalidInput("Attributes name is not valid UTF-8".to_string()))?;;
+            self.fs.set_x_attributes(path, name, value, flags, position).await;
+            Ok(())
+        } else {
+            Err(FuseError::Unsupported("setxattr".to_string()).into())
+        }
     }
 
     /// list extended attribute names. If size is too small, use [`ReplyXAttr::Size`] to return
     /// correct size. If size is enough, use [`ReplyXAttr::Data`] to send it, or return error.
     #[instrument(skip(self), err(level = Level::ERROR), ret(level = Level::DEBUG))]
     async fn listxattr(&self, req: Request, path: &OsStr, size: u32) -> FuseResult<ReplyXAttr> {
-        let path = Path::new(path);
-        let names = self.fs.list_x_attribute(path).await?;
-        let mut buf = Vec::new();
-        for name in &names {
-            buf.extend_from_slice(name.as_bytes());
-            buf.push(0);
+        if self.fs.use_xattributes() {
+            let path = Path::new(path);
+            let names = self.fs.list_x_attribute(path).await?;
+            let mut buf = Vec::new();
+            for name in &names {
+                buf.extend_from_slice(name.as_bytes());
+                buf.push(0);
+            }
+            if size == 0 {
+                return Ok(ReplyXAttr::Size(buf.len() as u32));
+            }
+            if size < buf.len() as u32 {
+                return Err(libc::ERANGE.into());
+            }
+            Ok(ReplyXAttr::Data(Bytes::from(buf)))
+        } else {
+            Err(FuseError::Unsupported("listxattr".to_string()).into())
         }
-        if size == 0 {
-            return Ok(ReplyXAttr::Size(buf.len() as u32));
-        }
-        if size < buf.len() as u32 {
-            return Err(libc::ERANGE.into());
-        }
-        Ok(ReplyXAttr::Data(Bytes::from(buf)))
     }
 
     /// remove an extended attribute.
     #[instrument(skip(self), err(level = Level::ERROR), ret(level = Level::DEBUG))]
     async fn removexattr(&self, req: Request, path: &OsStr, name: &OsStr) -> FuseResult<()> {
-        let path = Path::new(path);
-        let name = name.to_str().ok_or_else(|| FuseError::InvalidInput("Attributes name is not valid UTF-8".to_string()))?;;
-        let result = self.fs.remove_x_attributes(path, name).await?;
-        Ok(result)
+        if self.fs.use_xattributes() {
+            let path = Path::new(path);
+            let name = name.to_str().ok_or_else(|| FuseError::InvalidInput("Attributes name is not valid UTF-8".to_string()))?;;
+            let result = self.fs.remove_x_attributes(path, name).await?;
+            Ok(result)
+        } else {
+            Err(FuseError::Unsupported("removexattr".to_string()).into())
+        }
     }
 
     /// get filesystem statistics.
